feat(hitl): move models
